requirements.txt
Flask==2.0.2
paho-mqtt==1.5.1
python-dotenv==0.19.2
